fix: fixes rendering text with new lines

+++ source/Age.Rendering/Drawing/Style.cs
@@ -4,7 +4,7 @@ namespace Age.Rendering.Drawing;
 
 public record Style
 {
-    public int        FontSize { get; set; }
+    public ushort     FontSize { get; set; }
     public Point<int> Position { get; set; }
     public Size<uint> Size     { get; set; }
 }
